Add files via upload
Changes to REVEAL APP section :
-Turned off WinActivate
-Hard-coded ToolTip descriptions of audio device with mumble, Spotify, Chrome, VLC media player, and default programs

Changes to GetVolumeObject():
-Hard-coded which audio device to use based off of what app is being called.

I am unable to figure out how to look up what program is on what audio device, so i had to hard code them in. Perhaps a good solution to this is to make a configuration file (.ini) in which the user could specify what program should be paired with what device. Unfortanity, a lot of programs (Spotify, Chrome, ect.) do not let you set what output audio device you would like your audio to output to. So you will also need to follow this video: 
https://community.spotify.com/t5/Desktop-Linux-Windows-Web-Player/How-to-change-Spotify-output-device/td-p/1560358
https://github.com/audiorouterdev/audio-router
 
To set the output of your audio devices.